backend/requirements.txt
django
django-filter
django-redis
django-ninja
djangorestframework-simplejwt
djangorestframework

psycopg2